Reply to red0ne's comments

+++ x/shared/types/service_configs.go
@@ -95,7 +95,7 @@ func ValidateSupplierServiceConfigs(services []*SupplierServiceConfig) error {
 // ensuring that the sum of the revenue share percentages is 100.
 // NB: This function is unit tested via the supplier staking config tests.
 func ValidateServiceRevShare(revShareList []*ServiceRevenueShare) error {
-	revSharePercentageSum := uint32(0)
+	revSharePercentageSum := float64(0)
 
 	if len(revShareList) == 0 {
 		return ErrSharedInvalidRevShare.Wrap("no rev share configurations")
