+++ x/tokenomics/keeper/token_logic_modules_test.go
@@ -45,7 +45,7 @@ func TestProcessTokenLogicModules_TLMBurnEqualsMint_Valid(t *testing.T) {
 	// Test Parameters
 	appInitialStake := apptypes.DefaultMinStake.Amount.Mul(cosmosmath.NewInt(2))
 	supplierInitialStake := cosmosmath.NewInt(1000000)
-	supplierRevShareRatios := []float32{12.5, 37.5, 50}
+	supplierRevShareRatios := []float64{12.5, 37.5, 50}
 	globalComputeUnitsToTokensMultiplier := uint64(1)
 	serviceComputeUnitsPerRelay := uint64(1)
 	service := prepareTestService(serviceComputeUnitsPerRelay)
@@ -447,7 +447,7 @@ func TestProcessTokenLogicModules_TLMGlobalMint_Valid_MintDistributionCorrect(t 
 	propMint := cosmosmath.NewInt(int64(numTokensMinted * tokenomicsParams.MintAllocationPercentages.Proposer))
 	serviceOwnerMint := cosmosmath.NewInt(int64(numTokensMinted * tokenomicsParams.MintAllocationPercentages.SourceOwner))
 	appMint := cosmosmath.NewInt(int64(numTokensMinted * tokenomicsParams.MintAllocationPercentages.Application))
-	supplierMint := float32(numTokensMinted * tokenomicsParams.MintAllocationPercentages.Supplier)
+	supplierMint := float64(numTokensMinted * tokenomicsParams.MintAllocationPercentages.Supplier)
 
 	// Ensure the balance was increased to the appropriate amount.
 	require.Equal(t, daoBalanceBefore.Amount.Add(daoMint).Add(numTokensMintedInt), daoBalanceAfter.Amount)
@@ -459,7 +459,7 @@ func TestProcessTokenLogicModules_TLMGlobalMint_Valid_MintDistributionCorrect(t 
 		balanceBefore := supplierShareholderBalancesBefore[addr]
 		balanceAfter := supplierShareholderBalancesAfter[addr].Amount.Int64()
 		mintShare := int64(supplierMint * revShare.RevSharePercentage / 100)
-		rewardShare := int64(float32(numTokensClaimed) * revShare.RevSharePercentage / 100)
+		rewardShare := int64(float64(numTokensClaimed) * revShare.RevSharePercentage / 100)
 		balanceIncrease := cosmosmath.NewInt(mintShare + rewardShare)
 		expectedBalanceAfter := balanceBefore.Amount.Add(balanceIncrease).Int64()
 		// TODO_MAINNET(@red-0ne): Remove the InDelta check and use the exact amount once the floating point arithmetic is fixed
